equalsIgnoreCase_exercise
Add an if statement that checks to see if firstExample is equal to secondExample. If it is, print out "first is equal to second".

Add another if statement that checks if the firstExample is equal ignoring case to thirdExample. If it, is print out "first and third are the same ignoring case".